lambdas/recipe_scraper_lambda/requirements.txt
python-dotenv==1.0.0
httpx==0.25.2
recipe_scrapers==14.44.1
boto3==1.28.42 
s3fs==2023.10.0

# pandas==2.0.3 
# requests==2.31.0 